feat: add tfprovidertest package for identifying test coverage gaps in Terraform providers
- Implemented utility functions for converting between CamelCase and snake_case.
- Added functions to identify base class, sweeper, and migration files.
- Created logic to determine if a file should be excluded based on patterns.
- Developed functions to check for test function naming conventions.
- Implemented resource location formatting for enhanced issue reporting.
- Added functionality to extract resource names and attributes from schema definitions.
- Included checks for RequiresReplace plan modifier and ImportState method presence.
