+++ specs/todo_task.md
@@ -1,0 +1,175 @@
+The following markdown file details an analysis of the provided codebase, specifically focusing on the tfprovidertest linter and the associated validation CLI.
+
+It addresses your specific question regarding the viability of filename-based searches within this context and provides a comprehensive refactoring plan to eliminate redundancy and simplify the logic.
+
+# Codebase Analysis & Refactoring Recommendations
+
+## 1. Executive Summary
+
+The codebase currently suffers from a **"Split Brain" architecture**. There are two completely separate implementations of the logic used to identify Terraform resources and their associated tests:
+
+1.  **The Analyzer (`tfprovidertest.go`):** Uses robust AST (Abstract Syntax Tree) inspection to find structs with `Schema()` methods.
+2.  **The CLI (`cmd/validate/main.go`):** Uses a fragile, custom implementation that relies heavily on manual file walking and text parsing, ignoring the core analyzer logic.
+
+**Verdict on Filename-Based Searches:**
+
+- **For Resource Identification:** **No.** Relying on filenames (e.g., `resource_xyz.go`) to identify resources is error-prone. A file named `resource_utils.go` might contain helpers, not resources. Resources should be identified by **Code Structure** (implementing an interface or having a `Schema()` method).
+- **For Test Association:** **Yes.** Go testing conventions are file-based (`foo.go` -> `foo_test.go`). Using filenames to map resources to their test files is the standard, most reliable approach. The current code over-complicates this by trying to parse resource names out of function names (e.g., trying to extract `widget` from `TestAccWidget_basic`), which is brittle.
+
+---
+
+## 2. Issues Identified
+
+### Issue A: Logic Duplication (The "Split Brain")
+
+The `cmd/validate/main.go` file re-implements the discovery logic found in `tfprovidertest.go`.
+
+- **Risk:** If you update the analyzer logic to support a new Terraform plugin framework pattern, the CLI tool won't see it unless you duplicate the change.
+- **Evidence:** `main.go` manually parses files using `parser.ParseFile` and iterates over nodes, whereas `tfprovidertest.go` uses the `analysis.Pass` framework.
+
+### Issue B: Monolithic Source File
+
+`tfprovidertest.go` is over 1,000 lines long. It contains:
+
+1.  Configuration (`Settings`)
+2.  Data structures (`ResourceRegistry`)
+3.  AST Parsing logic (`parseResources`)
+4.  Linter analysis logic (`runBasicTestAnalyzer`)
+5.  String manipulation utilities
+
+This makes unit testing difficult and readability poor.
+
+### Issue C: Over-Engineered Test Matching
+
+The function `extractResourceNameFromTestFunc` in `tfprovidertest.go` is excessively complex. It attempts to support every possible naming variation (CamelCase, snake_case, truncated prefixes) to guess which resource a test belongs to.
+
+- **Impact:** This results in false positives and maintenance overhead.
+- **Better Approach:** Trust the file pairing. If `TestAcc_Basic` is inside `resource_widget_test.go`, it belongs to `resource_widget.go`.
+
+### Issue D: The Python Patch Script
+
+The presence of `update_tfprovidertest.py` indicates that `tfprovidertest.go` is either incomplete or being patched during a build process. This is a "code smell." The logic added by the script (custom patterns, exclusion logic) should be native to the Go codebase.
+
+---
+
+## 3. Recommended Solution
+
+### Step 1: Delete Custom Logic in CLI
+
+The `cmd/validate/main.go` should be a thin wrapper that invokes the analyzers defined in the package. It should not perform parsing itself.
+
+**Recommended `cmd/validate/main.go`:**
+
+```go
+package main
+
+import (
+	"github.com/golangci/plugin-module-register/register"
+	"golang.org/x/tools/go/analysis/multichecker"
+    // Import the package containing the analyzers
+	"path/to/hashi-demo-lab/tfprovidertest"
+)
+
+func main() {
+    // Determine which analyzers to run based on settings or flags
+    // This uses the standard Go analysis runner which handles
+    // flags, file loading, and parsing automatically.
+
+    plugin, _ := tfprovidertest.New(nil)
+    analyzers, _ := plugin.BuildAnalyzers()
+
+	multichecker.Main(analyzers...)
+}
+```
+
+### Step 2: Modularize the Package
+
+Break `tfprovidertest.go` into focused files:
+
+- `analyzer.go`: Contains the `analysis.Analyzer` definitions and `run*` functions.
+- `registry.go`: Contains `ResourceRegistry`, `ResourceInfo`, and `TestFileInfo` structs.
+- `parser.go`: Contains `parseResources`, `parseTestFile`, and AST inspection logic.
+- `utils.go`: String helpers (`toSnakeCase`, `toTitleCase`).
+- `settings.go`: Configuration structs.
+
+### Step 3: Simplify Test Association
+
+Replace the complex function name parsing with a robust file-correlation strategy.
+
+**Proposed Logic:**
+
+1. **Identify Resources (AST):** Find all structs that implement the Resource interface.
+2. **Identify Test Files (File System):** Find all `*_test.go` files.
+3. **Associate (Path):**
+   - If `resource_widget.go` defines `WidgetResource`
+   - And `resource_widget_test.go` exists in the same package
+   - Then `resource_widget_test.go` is the test file for `WidgetResource`
+4. **Verify (Function Content):** Inside `resource_widget_test.go`, look for any function starting with `TestAcc`. We don't need to parse the function name to know it belongs to `WidgetResource` because the file tells us that.
+
+### Step 4: Eliminate Python Script
+
+Manually apply the changes from `update_tfprovidertest.py` into the Go source code.
+
+- Add the `shouldExcludeFile` function to `utils.go`.
+- Update the `parseTestFile` signature in `parser.go`.
+- Delete `update_tfprovidertest.py`.
+
+---
+
+## 4. Simplified Code Example (Parser)
+
+Here is how the simplified "File-First" association logic should look. This replaces ~200 lines of regex/string guessing code.
+
+```go
+// In parser.go
+
+// Parse the package to find resources and tests
+func buildRegistry(pass *analysis.Pass) *ResourceRegistry {
+    reg := NewResourceRegistry()
+
+    // 1. Scan for Resources (Typed based)
+    for _, file := range pass.Files {
+        // ... standard AST inspection to find Schema() or interface implementation ...
+        // Register resource with its defined filename
+    }
+
+    // 2. Scan for Test Files (File based)
+    // Go analysis provides access to test files in the package if run with -t
+    for _, file := range pass.Files {
+        filename := pass.Fset.Position(file.Pos()).Filename
+        if !strings.HasSuffix(filename, "_test.go") {
+            continue
+        }
+
+        // Simple Association: "resource_foo_test.go" -> "resource_foo.go"
+        expectedResourceFile := strings.TrimSuffix(filename, "_test.go") + ".go"
+
+        // Find which resource is defined in expectedResourceFile
+        resource := reg.GetResourceByFile(expectedResourceFile)
+
+        if resource != nil {
+            // Found it! This test file belongs to that resource.
+            // Now just check if it has valid TestAcc functions.
+            hasTests := containsTestAcc(file)
+            reg.RegisterTestCoverage(resource.Name, hasTests)
+        }
+    }
+    return reg
+}
+
+func containsTestAcc(file *ast.File) bool {
+    hasTest := false
+    ast.Inspect(file, func(n ast.Node) bool {
+        if fn, ok := n.(*ast.FuncDecl); ok {
+            if strings.HasPrefix(fn.Name.Name, "TestAcc") {
+                hasTest = true
+                return false
+            }
+        }
+        return true
+    })
+    return hasTest
+}
+```
+
+This approach is significantly more robust. It respects the standard Go convention that tests live next to the code they test, rather than relying on developers naming their test functions perfectly.
