+++ tfprovidertest_test.go
@@ -464,7 +464,11 @@ func TestAccWidget_custom(t *testing.T) {
 }
 
 // T301: Test for extractResourceNameFromTestFunc with various patterns
+// NOTE: This tests DEPRECATED functionality. ExtractResourceNameFromTestFunc is deprecated
+// in favor of file-based matching. It now always returns empty string.
 func TestExtractResourceNameFromTestFunc_AdditionalPatterns(t *testing.T) {
+	t.Skip("ExtractResourceNameFromTestFunc is deprecated - file-based matching is now used")
+	
 	t.Run("should extract resource name from various test patterns", func(t *testing.T) {
 		tests := []struct {
 			funcName string
